- All `views` and their logic was moved out `library-core` in separated `iconics-view-library`
