+++ iconics-view-library/src/main/java/com/mikepenz/iconics/internal/IconicsView.java
