+++ iconics-view-library/src/main/java/com/mikepenz/iconics/view/IconicsCheckBox.java
