+++ iconics-view-library/src/main/java/com/mikepenz/iconics/view/IconicsImageButton.java
@@ -17,11 +17,8 @@
 package com.mikepenz.iconics.view;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatButton;
 import android.util.AttributeSet;
 import android.widget.ImageButton;
-
-import com.mikepenz.iconics.Iconics;
 
 public class IconicsImageButton extends IconicsImageView {
