+++ iconics-view-library/src/main/java/com/mikepenz/iconics/view/IconicsImageView.java
@@ -28,9 +28,8 @@ import android.support.v7.widget.AppCompatImageView;
 import android.util.AttributeSet;
 
 import com.mikepenz.iconics.IconicsDrawable;
-import com.mikepenz.iconics.core.R;
-import com.mikepenz.iconics.internal.AttributeSetReader;
 import com.mikepenz.iconics.internal.IconBundle;
+import com.mikepenz.iconics.internal.IconicsCoreAttrsReader;
 import com.mikepenz.iconics.internal.IconicsView;
 import com.mikepenz.iconics.typeface.IIcon;
 import com.mikepenz.iconics.utils.Utils;
@@ -78,7 +77,7 @@ public class IconicsImageView extends AppCompatImageView implements IconicsView 
         final TypedArray a = context.obtainStyledAttributes(attrs, R.styleable.IconicsImageView, defStyle, 0);
     
         //set the color even if we had no image yet
-        AttributeSetReader.readIconicsImageView(a, mIconBundle);
+        IconicsCoreAttrsReader.readIconicsImageView(a, mIconBundle);
     
         //recycle the typedArray
         a.recycle();
