+++ library-core/src/main/res/values/strings.xml
@@ -1,12 +1,3 @@
 <resources>
     <string name="app_name">library</string>
-
-    <string name="default_icon"> </string>
-    <integer name="default_color">0</integer>
-    <integer name="default_size">-1</integer>
-    <integer name="default_padding">-1</integer>
-    <integer name="default_contour_color">0</integer>
-    <integer name="default_contour_width">-1</integer>
-    <integer name="default_background_color">0</integer>
-    <integer name="default_corner_radius">-1</integer>
 </resources>
